Moved all receiver/prometheusreceiver/targetallocator to receiver/prometheusreceiver/internel
