--- receiver/prometheusreceiver/targetallocator/testdata/config.yaml
@@ -1,8 +1,0 @@
-target_allocator:
-  endpoint: http://localhost:8080
-  timeout: 5s
-  tls:
-    cert_file: "client.crt"
-    key_file: "client.key"
-  interval: 30s
-  collector_id: collector-1
